refac: Roll back to a single general Configuration (#404)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Roll back from algorithm-specific configurations
(`N2VConfiguration`, `N2NConfiguration` and `CAREConfiguration`) to a
single `Configuration` .


### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

While not an issue, the split between different algorithm-specific
configuration is not necessary since
https://github.com/CAREamics/careamics/pull/365 removed N2V data
configuration. Algorithm-specific configurations add a level of
complexity that is now unnecessary.


### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

All configurations are now instantiated by a single `Configuration`
class.

### Implementation - how did you implement the changes?

<!-- How did you solve the issue technically? Explain why you chose this
approach and
provide code examples if applicable (e.g. change in the API for users).
-->

The algorithm-specific configurations are gone, and the choice between
algorithms is now done via discrimination of the algorithm configuration
themselves.

## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### New features or files

<!-- List new features or files added. -->
- `Configuration` now allows a range of algorithm configurations and
calls their BMZ-related algorithm metadata functions.

### Modified features or files

<!-- List important modified features or files. -->
- All the BMZ-related algorithm metadata have moved to algorithm
configurations.
- All tests have been updated
- Every call to `create_configuration` has been updated to
`Configuration(**kwargs)`

### Removed features or files

<!-- List removed features or files. -->
- `N2VConfiguration`, `N2NConfiguration` and `CAREConfiguration` were
deleted
- `create_configuration` was deleted

## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->


## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

- Resolves https://github.com/CAREamics/careamics/issues/367
- Resolves https://github.com/CAREamics/careamics/issues/384

## Breaking changes

<!-- Describe any breaking changes introduced by this PR. -->

Any code that uses the algorithm-specific configurations.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)