Enforce `shuffling` in convenience functions (#413)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Prevent users from running into errors as soon as they add
training dataloader parameters.

I ran into this error while fixing the docs: if one wishes to change
`num_workers` in the train dataloader, then an error would be raised
(Pydantic validation) if the `shuffle` parameter is also absent.

This PR enforces the presence of the `shuffle` parameter for users.


### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

The current mindset of the configuration is the following:
- Pydantic configuration do not change the user input and the sole
responsibility is coherence and clarity of errors (at least those we
write).
- Convenience functions are the entry point for users and these should
minimize the potential for misconfigurations (and therefore validation
errors).

The current behaviour leads to easy to stumble upon errors: any
parameter but `shuffle` passed to the `train_dataloader_params`
parameter of the convenience functions will lead to an error thrown at
the user.

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

This PR enforces the addition of `shuffle` to the
`train_dataloader_params` if it was not passed by users to the
convenience functions.

### Implementation - how did you implement the changes?

<!-- How did you solve the issue technically? Explain why you chose this
approach and
provide code examples if applicable (e.g. change in the API for users).
-->

The convenience functions call `_create_data_configuration`, which now
test for the presence of `shuffle` and adds it if absent.

## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### Modified features or files

<!-- List important modified features or files. -->
- `_create_data_configuration` checks for presence in
`train_dataloader_params` and add `"shuffle" : True` if absent
- Updated the description of the parameter to mention the necessity to
have the `shuffle` parameter.
- Added a test to check the behaviour.


## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->

Locally and via the tests.

## Breaking changes

<!-- Describe any breaking changes introduced by this PR. -->

No breaking changes I can think of.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [x] PR to the documentation exists (for bug fixes / features)