Feature(Next-Gen Dataset): Improved patching strategy classes (#439)
## Description

> [!NOTE]  
> **tldr**: This PR addresses the shortcomings of the original draft
implementation of the `PatchSpecsGenerator` classes, that were
discovered after trying to use them in the prototype dataset in #420.
This includes having a random patching strategy that will change each
epoch. They have also been renamed to `PatchingStrategy` classes.

### Background - why do we need this PR?

The previous `PatchSpecsGenerator` classes weren't working very well
when we tried to use them in the dataset prototype in #420, and the name
is also hard to say 😅. The main problem was that the random patches
wouldn't change each epoch.

### Overview - what changed?

Added a `RandomPatchingStrategy`, `FixedRandomPatchingStrategy` and
`SequentialPatchingStrategy`.

Renamed `PatchSpecsGenerator` to `PatchingStrategy`. Now the classes
should also be initialised with the datashapes and the patch size, since
for a given `CAREamicsDataset` instance we don't expect these to change.
The `PatchingStrategy` protocol dictates that a concrete implementation
should have the methods `n_patches` that should return the total number
of patches, and a `get_patch_spec` method that takes an index and will
return a `PatchSpecs` dict.

EDIT: `CAREamicsDataset` now takes a `mode` argument to determine
whether it is training, validating or predicting. The `mode` will
determine the patching strategy: `RandomPatching` for training,
`FixedRandomPatching` for validating and `TilingStrategy` for
predicting. (`TilingStrategy` is implemented in an upcoming PR).

### Implementation - how did you implement the changes?

The `RandomPatchingStrategy` creates index bins to work out what
image_stack and sample a given patch index maps to, and then generates a
random patch. This is to ensure the random patching still produces a
fixed known number of patches for each image_stack and sample within the
image_stack, but is then random spatial location within the single
sample.

~~In the dataset the patching strategy should be a attribute and during
`__getitem__`, `get_patch_spec` will be called to index the data.~~
EDIT: In the dataset the patching strategy is an attribute and during
`__getitem__`, `get_patch_spec` is called to index the data.

See sketch code below:

```python

class CAREamicsDataset:

    def __init__(self, ...):
        patch_extractor = create_patch_extractor(...)
        patching_strategy = RandomPatchingStrategy(...) #maybe we want to make this swappable

    def __len__(self):
        return patching_strategy.n_patches()

    def __getitem__(self, index: int):
        patch_spec = self.patching_strategy.get_patch_spec(index)
        patch = patch_extractor.extract_patch(**patch_spec)
        return patch
```

## Changes Made

### New features or files

- `RandomPatchingStrategy` - patches will change each epoch
- `FixedRandomPatchingStrategy` - patches are fixed each epoch
- placeholder `SequentialPatchingStrategy`

### Modified features or files

- A couple organisational changes, like moving the `PatchSpecs`
definition.
- Exported all the notebooks to scripts because mypy isn't set up to
look for errors in them
- Updated `demo_patch_extractor` to use the new `RandomPatchingStrategy`
- EDIT: `CAREamicsDataset.__init__`
- EDIT: `CAREamicsDataset.__getitem__`

### Removed features or files

- remove the method `extract_patches` from `PatchExtractor` because
extracting multiple patches isn't needed anymore

## How has this been tested?

Added tests

## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

- #425 mostly solved apart from `WholeSamplePatchingStrategy`

## Additional Notes and Examples

I think I should wait for #420 to be merged and I will implement
integration with dataset class

I will address tiling for prediction where we need the overlap coords in
an up coming PR

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)