Feature(Next-Gen Dataset): Prediction patching (tiling) (#444)
## Description

> [!NOTE]  
> **tldr**: Introducing `PatchingStrategies` for prediction. These are a
`TilingStrategy` that will output a `TileSpecs` object, which inherits
from the `PatchSpecs` type but includes the additional fields
`crop_coords`, `crop_size` and `stitch_coords`; the second patching
strategy is the `WholeSampleStrategy`, this is for prediction without
tiling and will select a whole sample.


### Background - why do we need this PR?

To be able to stitch predicted tiles back together we need some
additional information that is not included in the current `PatchSpecs`
type. These are the:
- `crop_coords`: where will the tile be cropped from to deal with
overlaps, this is the top-left corner, relative to the tile.
- `crop_size`: the size of the crop. So the tile will be cropped to
`crop_coords: crop_coords + crop_size`
- `stitch_coords`: where the cropped tiles will be placed in the
stitched image, top-left corner, relative to the image.
So this PR introduces the `TileSpecs` `TypedDict` to store this
information, and a `TilingStrategy` class to produce it.

### Overview - what changed?

In the `CAREamicsDataset` if the `mode` is set to predicting, it will
chose the `TilingStrategy` if both the `tile_size` and `tile_overlaps`
are not `None` in the `InferenceConfig`, otherwise it will choose the
`WholeSampleStrategy`.

Also a `Imageregions_to_tileinfos` function has been introduced to
convert the new `ImageRegionData` to `TileInformation` objects so the
existing `stitch_prediction` functions can be used as they are (but they
should be updated in the future, once the dataset is complete).

### Implementation - how did you implement the changes?

The implementation of the `TilingStrategy` is largely adapted from the
existing `extract_tiles` function.

## Changes Made

### New features or files

- `TileSpecs` typed dict
- `TilingStrategy` patching strategy
- `WholeSamplePatchingStrategy` patching strategy
- `imageregions_to_tileinfos` function

### Modified features or files

- `CAREamicsDataset` patching strategy initialisation

## How has this been tested?

- Added test to make sure the stitched array is equal to the original
array, when using tiles generated from `TilingStrategy` and the existing
CAREamics `stitch_prediction` functions, in both 2D and 3D.
- Added `TilingStrategy` and `WholeSamplePatchingStrategy` to the
`test_all_strategies.py` tests.


## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

- Resolves #425 

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

---------

Co-authored-by: Igor Zubarev <zubarev.ia@gmail.com>