Fix: artefacts from two incompatible PRs being merged into main (#455)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Minor incompatibility with two merged PRs


#444 Introduced a new test that required the `create_patch_extractor`
function. But #449 removed this function in favour of individual
functions to create each type of patch extractor. I replaced the
function with the relevant `create_array_extractor`.

Somehow import of `Generic` from typing was missing in the main dataset
module.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>