Next-Gen dataset: dataset statistics (#466)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: <!-- Write a one sentence summary. -->
This PR introduces statistics calculation for the next-gen dataset.

### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->
To match the functionality of the old datasets, we need to be able to
calculate the dataset statistics before training to apply the
appropriate normalization.

### Implementation - how did you implement the changes?

<!-- How did you solve the issue technically? Explain why you chose this
approach and
provide code examples if applicable (e.g. change in the API for users).
-->

I copied the statistics calculation from the `iterable_dataset.py`. If
the statistics are not defined in the config, they'll be calculated with
`WelfordStatistics` by iterating through all the patches in the given
image stack.

Also, following a todo in
https://github.com/CAREamics/careamics/pull/461, I added a check in
`normalization.py` to check that target stats are present if the target
is not None and a relevant test.

## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### Modified features or files

<!-- List important modified features or files. -->
- added statistics calculation to `CareamicsDataset` 
- added stats emptiness check to `normalization.py` 

## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->
I removed the stats calculation from the demo notebooks, ran the
training, and made sure that the final metrics roughly match (explained
below).

## Additional Notes and Examples

<!-- Provide any additional information that will help reviewers
understand the
changes. This can be links to documentations, forum posts, past
discussions etc. -->

A discussion point for the future: while testing the notebooks, I
learned that slight differences in statistics can significantly
influence the resulting metrics for some datasets.

Example - I trained N2V on bsd64, setting the mean and std to
`data.mean()` and `data.std()`, and got a metric slightly lower than
reported previously. It turned out that in the old dataset, the
statistics were calculated on the pre-computed patches, while I was
computing them on the whole data, and the difference was significant
enough to influence the final metric.

In this PR, the statistics are calculated using the selected patching
strategy, so for training, it will be `RandomPatchingStrategy`. Random
patching will influence the statistics a bit. We need to test how
critical it is for real microscopy datasets and how we can address that.

---

**Please ensure your PR meets the following requirements:**

- [ ] Code builds and passes tests locally, including doctests
- [ ] New tests have been added (for bug fixes/features)
- [ ] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)