fix: Fix pre-commit by adding missing docstrings (#484)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Pre-commit currently fails because of missing dosctrings,
this PR adds the missing docs.

> [!IMPORTANT]  
> - pre-commit will still fail because of a `mypy` error that will be
fixed in https://github.com/CAREamics/careamics/pull/472.
> - I excluded `data_module.py` from numpydoc because it tries to
enforce documenting the overloads, which is incompatible with ruff.
> - I might not have adhered strictly at our conventions and the
docstrings will have to be reviewed once we document the dataset on the
website. We will anyway need to do a complete review of the docstrings
because the formatting is failing at times on the website.

### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

Missing docstrings NG dataset files are causing pre-commit to fail on
all PRs.

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

Added docstrings.

### Implementation - how did you implement the changes?

<!-- How did you solve the issue technically? Explain why you chose this
approach and
provide code examples if applicable (e.g. change in the API for users).
-->


## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### Modified features or files

<!-- List important modified features or files. -->
- `src.careamics.lightning.dataset_ng.*`

## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->

N/A

## Breaking changes

<!-- Describe any breaking changes introduced by this PR. -->

Potential conflicts with
https://github.com/CAREamics/careamics/pull/472.

## Additional Notes and Examples

<!-- Provide any additional information that will help reviewers
understand the
changes. This can be links to documentations, forum posts, past
discussions etc. -->

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [ ] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>