Refac(Next-Gen): Fold patching strategies into a subconfiguration of DataConfig (#486)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Unify `InferenceConfig` and `DataConfig` into a single
configuration for the Next-Gen Dataset, by moving all patching strategy
related parameters in the configuration into a sub-configuration.

### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

CAREamics rely on two different data configuration: one for
training/validation, the other one for inference. This was not an issue
when both steps were handled by different datasets. The new Next-Gen
Dataset will handle all datasets in CAREamics, and the code would be
greatly simplified if both steps were also handled by a single data
configuration.

This PR offers a potential solution to handling both train/val and test
in a single data configuration.

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

A new `NGDataConfig` class has been added that handles all the
parameters necessary for the various patching strategy and dataloader
parameters. The subconfiguration `PatchingModel` holds the parameters
related to the patching strategies.

The `dataset` and `data_module` have been updated to only take a single
data config as input.

### Implementation - how did you implement the changes?

<!-- How did you solve the issue technically? Explain why you chose this
approach and
provide code examples if applicable (e.g. change in the API for users).
-->

Here, the previous `patch_size` parameter has been folded into a
subconfiguration for the strategies, and also corresponds to the
`tile_size`. The patching strategy is chosen by setting the `name`
parameter used as discriminator.

Each patching strategy has a Pydantic model, with some validation if
applicable. I chose inheritance to avoid duplicating code (e.g.
validation of parameters).

## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### New features or files

<!-- List new features or files added. -->
- New `NGDataConfig` class, with `seed`, `patching_strategy` and
`test_dataloader_params` new parameters
- Various patching strategy have their own Pydantic model

### Modified features or files

<!-- List important modified features or files. -->
- Dataset and data module now take `NGDataConfig` as sole configuration.

## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->

Tests locally, ran the local NG Dataset notebooks and NG Dataset BSD68.

## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

- Resolves #

## Breaking changes

<!-- Describe any breaking changes introduced by this PR. -->

## Additional Notes and Examples

<!-- Provide any additional information that will help reviewers
understand the
changes. This can be links to documentations, forum posts, past
discussions etc. -->

Note that an additional PR would be necessary to integrate this
`NGDatasetConfig` into the main configuration. I made the choice to not
introduce breaking change for the rest of the code.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

---------

Co-authored-by: Melisande Croft <63270704+melisande-c@users.noreply.github.com>