chore: CZI format as optional dependency (#492)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: CZI deps might be difficult to install for macOS users, this
PR makes it optional.


### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

macOS users need to install `cmake` in order to compile the `pylibCZIrw`
package, which may lead to many users encountering issues.

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

The `CZIImageStack` now attempts to load `pylibCZIrw` and fails upon
instantiation if it is not present, with a user friendly usage.

CAREamics has now a `czi` optional dependency group.

### Implementation - how did you implement the changes?

<!-- How did you solve the issue technically? Explain why you chose this
approach and
provide code examples if applicable (e.g. change in the API for users).
-->

``` python
try:
    from pylibCZIrw.czi import CziReader, Rectangle, open_czi

    pyczi_available = True
except ImportError:
    pyczi_available = False

if TYPE_CHECKING:
    from pylibCZIrw.czi import CziReader, Rectangle, open_czi
```


## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### New features or files

<!-- List new features or files added. -->
- Test testing the failure of the `CziImageStack` in the absence of the
czi format package.

### Modified features or files

<!-- List important modified features or files. -->
- `CziImageStack`
- corresponding tests
- `pyproject.toml`

## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->

Tests.

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)