chore: Drop support for 3.9 (#485)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Update requirements and python syntax to >= py3.10, dropping
py3.9.

[Python 3.9](https://devguide.python.org/versions/) is being dropped
this year, and it is time to drop 3.9. This PR updates the syntax of the
code base via `ruff`, and remove the 3.9 from various places (e.g., BMZ,
CI, metadata).

### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

3.9 is being dropped and won't be supported in the future.

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

Mostly requirements in CI, metadata, and code syntax throughout the
codebase.

### Implementation - how did you implement the changes?

<!-- How did you solve the issue technically? Explain why you chose this
approach and
provide code examples if applicable (e.g. change in the API for users).
-->

I updated the requirements in the pre-commit configuration and
pyproject, then ran `ruff` to automatically update the syntax in the
code. I then fixed manually the various `(str, Path)` -> `str | Path`
occurrences.

Finally, I removed 3.9 CIs.

## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### New features or files

<!-- List new features or files added. -->

Changes have happened throughout the code thanks to `ruff`.

## How has this been tested?

<!-- Describe the tests that you ran to verify your changes. This can be
a short
description of the tests added to the PR or code snippet to reproduce
the change
in behaviour. -->

Running tests via CI here.

## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

Note that it will allow https://github.com/CAREamics/careamics/pull/476
to have the CI pass, and will interact with all other open PRs because
the changes are many.


**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [ ] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>